Compile with no implicit returns.
I'm prototyping using this module as a plugin in https://github.com/bazelbuild/rules_typescript

As part of pulling it into google's internal repo for the prototyping, it has to compile with google's internal tsconfig, which includes noImplicitReturns. I'm not sure that it's an improvement in legibility, so if you don't want to merge this that's fine, we can use a patch set on top.