Add the ability to emit index entries.
I'm prototyping a Kythe TypeScript indexer plugin and things are fitting together pretty well.

This starts by adding a method on the analyzer that will yield all of the information that the analyzer has determined about a file. I started with element and attribute references as they seem like the most valuable. What else could we emit here?