Update `vscode-styled-components` from a git dep to a tarball to avoid install scripts.
The install scripts for the previous version attempt to fetch a URL that now
consistently 404's. However, this dependency is only needed for a couple of
files that are checked directly into the repo, so there's no need to actually
build the package. The commit that was used before isn't available as a release
on GitHub but the latest version has the same files at the same location (with
seemingly minor updates), so I've also started referencing the latest version.

FWIW, `vscode-styled-components` is still not available as an npm package.