feat: update web-component-analyzer and typescript
typescript 5.x removed a few deprecated methods, one of which we
depended on inside WCA. That has since been fixed over in the WCA repo.

This change is to now bump our dependency of WCA so we pull the fix in.

Meanwhile, typescript has also been updated to match the same semver as
what WCA depends on (otherwise we get type mismatches when passing
`SourceFile` objects around).